+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- API NSW Liquor Licence Element
 - Section element header to tabindex
 - Focusing of first element on a page when Next button is focused and used to nav forwards
 - `role` attribute to `<DropdownMenu />` component for accessibility